requirements.txt
streamlit
streamlit-folium
geopandas
pandas
folium
google-cloud-bigquery
google-auth
rapidfuzz
haversine
openpyxl
db-dtypes
xlrd
xlsxwriter